requirements.txt
fastapi
uvicorn 
pytesseract
pdfminer.six
python-docx
sentence-transformers
pinecone-client
spacy
rank_bm25
torch
transformers
opencv-python
unstructured
tiktoken
boto3
textractor
numpy
sckit-learn